apps/orchestrator/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-aiplatform==1.38.1
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
tweepy==4.14.0
praw==7.7.1
newsapi-python==0.2.6
googlemaps==4.10.0
firebase-admin==6.2.0
textblob==0.17.1 